+++ 1-exchanges/flexiswap/src/transporter/mod.rs
@@ -1,4 +1,4 @@
 mod authentication;
-mod transporter;
+mod blueprint;
 mod decoder;
 mod voucher;
